djangoBackend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
sqlparse==0.5.5
